include RequestStore clearing into specs
